TreeEditor: added 'activated' trait to notify a node being activated.
The 'activated' trait Event is assigned a node when a node is
double-clicked or enter key is pressed on a node.
Previously, the dclicked was used for such behaviour, which was also
fired on Enter key press in wx backend. This is now fixed and dclick
only fires on mouse double-click on a node in both qt and wx.