FIX: size policy of non-resizable items and labels
In the Qt backend, resizing a checkbox element with a label on the
right resizes the checkbox, even though it cannot be stretched horizontally.

The final effect is that the label remains attached to the right margin,
with a big gap between it and the checkbox. In this case, the label
should be made resizable instead.

On the other hand, a text element should stretch.

This commit adds tests and fixes for this behavior.