FIX: visible_when conditions make dialog too large
The size of a dialog that contains elements that are activated
by "visible_when" can end up being the *sum* of the sizes of the
elements, even though the elements are mutually exclusive (e.g.,
a typical case is a dropbox that lets you select different cases).
The expected behavior is that the size of the dialog should be at most
the size of the largest combination of elements.

This commit contains tests for this and fixes the bug in the
Qt backend (but not in wx).