FIX TreeEditor should release listeners to root node object
With Qt, when hide_root==True, to root node is assigned to
the tree's "invisibleRootItem", and listeners are added
as usual on the root node object.

However, when disposing the TreeEditor, the listener was
not released. This commit fixes that. This issue caused
a traceback when using Mayavi's "engine view", then closing
the "engine view" before exiting the application.