Fixes issue where empty nodes were not changing when adding children.
When going from having no children to having children, nodes were not being
updated to indicate that they could be expanded.  The tree editor handles this
by adding a dummy child to the unexpanded node so that Qt will flag it as
expandable.  This was not happening in the handlers for replacing all children
or updating children for a node in the case where they were not expanded and
started with no children.