Ensure other libraries know which toolkit to try.
This is a fix for #258.  In the case that a toolkit has not yet been selected,
when we try to select a toolkit we temporarily set the ETSConfig.toolkit to
the toolkit we are trying to import so ensure that other libraries
(in particular, pyface) don't try to second-guess our current choice.

The current implementation is sub-optimal in that it assumes things about the
implementation of ETSConfig which it shouldn't.