Fix wx capture mouse issue.
wxPython 3 is more strict about capturing the mouse, and no longer allows a widget to capture it if it already holds the capture, and it's possible for the capture to be stolen by another widget.  So trying to track the state of the capture in a local attribute can sometimes have holes, resulting in assertion errors from wx.