Fixes exception occurring during disposal of ListStrEditor
Solves issue #269. The root cause was that the adapter was not able to deal with
None objects. This is already performed in the TabularAdapter, and was missing in
the ListStrAdapter. The result was that ListStrModel.rowCount was raising exception
because at deinitialization, the editor trait object can be None. It is sensible in
this case that the adapter simply returns zero.