Do not destroy children widgets on ui.dispose()
Qt automatically manages destruction of child widgets when the
parent is destroyed. We do not have to explicitly iterate over
child widgets and delete them. In fact it causes occasional
segfaults on exit when the deleted child is attempted to be
deleted again when the parent is destroyed on exit.