Resolve spin control editing test.
The issue which test_wx_spin_control_editing_does_not_update tries to
cover is no longer an issue under Wx 3.0+, ie. text change events are
emitted in the same way as spin button events so we don't need to
explicitly listen for them, and closing the window causes loss of
focus which causes the expected update.  Hence we can skip the test
under WxPython 3+