Fix IconButton with a clickable area too small.
If we set the maximum size using the icon size, the clickable area
ends up being too small to click (tested on OSX with FileEditor,
RangeEditor, ListEditor). If we don't set the maximum size, the
button takes up too much space. This sizeHint logic is similar
to how QToolButton.sizeHint is implemented.