Refactored the RangeEditor backend for Qt and Wx so that their behavior becomes more consistent. (Closes issue #1771)
Updated the key_click_text_entry in wx/_interaction_helpers.py to respond to "Right" or "Left" interaction. (Closes issue #1772)

Replaced the use of possibly insecure eval function with the safer ast.literal_eval in RangeEditor. (See issue #1773)

Fixed a bug on wx: the text style RangeEditor does not raise an error whenever the textbox is empty,
even if enter has not been pressed anymore.

Update SimpleSpinEditor to allow floats and be able to specify the step of the spin control. Also added the possibility to increase the step value for the spinner by the use of the modifier keys "Shift", "Ctrl" or "Alt". They will increase the step value for the spinner by a factor of 2, 10 and 100, respectively. If a combination of "Shift", "Ctrl", "Alt" -keys are pressed the step increment will be the product of their factors. (Closes issue #1776)

Changed the default low and high attribute of RangeEditor from 0, 1 to undefined or None, so that
the limits are from -infinite to +infinite by default. (Closes issue #1775)

Made the behavior of increase_range and decrease_range methods of LargeRangeEditor more userfriendly. (Closes issue #1777)

Fixed so that the test_range_editor_text.py does not freeze due to a pop up error dialog (Closes issue #1778)

Changed the default behavior of mode attribute in the RangeEditor. Now the the 'View' conforms to the view explicitly set by the mode attribute. (Closes issue #1779)

Updated the mouse_click function in wx/_interaction_helpers.py to set focus to current control (Closes issue #1780)

Registered interaction helper for SimpleSpinEditor for both Qt and Wx: (traitsui/testing/tester/_ui_tester_registry/qt4/_traitsui/range_editor.py,
traitsui/testing/tester/_ui_tester_registry/wx/_traitsui/range_editor.py)