Use two-phase HeadingText creation from Pyface 7.4.0.
This removes a deprecation warning coming from a change in Pyface behaviour.
Additionally it adds proper disposal of the HeadingText Widget to the
TitleEditor.  It would be nice to do the same for the ui_panel code, but there
is no available hook and deeper architectural issues with the code.