Fix test suite hang on wx (#1903)
This PR adds a skip for a test that hangs on wxPython. On my machine, the wxPython test run now executes to completion on my machine. (It has many failures and an occasional segfault, but that seems like a slightly better state than a hang - we at least get to find out how many tests failed.)

Closes #1901.