Update etstool for 3.8 default (#1962)
This updates etstool to allow install of both 3.6 and 3.8 EDM runtimes. This is fairly minimalist in approach and it doesn't attempt to make things work for CI in Python 3.8 (that will be a separate PR).

This also fixes #1968 by shipping an edm.yaml file and updates some of the CI to account for the move of `ubuntu-latest` to `ubuntu-22.04` which doesn't easily provide Python 3.6 runtimes.