Remove `merge_undo` from undo code. (#1999)
Fixes #1515.  Also removes a no-longer used `history` trait.

This completes the transition of TraitsUI undo/redo to share the same basis as the Pyface/Apptools undo/redo.

There is still an issue in #1060 that is unresolved which is about refactoring TraitsUI to be more command-stack based, rather than building the command stack reactively.