Replace 'fast_ui' dispatch with 'ui' dispatch (#2009)
The `'fast_ui'` dispatch mechanism is identical to the `'ui'` dispatch
mechanism. It's not used by the `fast_ui` name anywhere in ETS outside
the the TraitsUI `TreeNode` implementation.

This PR replaces use of `fast_ui` with `ui`; this should allow us to
deprecated and remove the `fast_ui` dispatch method in Traits at some
point in the future.